sanitized filename fix, finished nfo files

+++ app/src/main/java/ryzen/ownitall/Sync.java
@@ -87,7 +87,7 @@ public class Sync {
         LocalDate currentDate = LocalDate.now();
         String folderName = currentDate.format(DateTimeFormatter.ofPattern("dd-MM-yyyy"));
         File archiveFolder = new File(this.dataFolder, folderName);
-        if (archiveFolder.exists() && !userInput) {
+        if (archiveFolder.exists() && userInput) {
             try {
                 System.out.println(
                         "You are about to overwrite the contents of the folder: " + archiveFolder.getAbsolutePath());
@@ -134,8 +134,18 @@ public class Sync {
                 if (file.isFile()) {
                     File destFile = new File(this.dataFolder, file.getName());
                     if (destFile.exists()) {
-                        destFile.delete();
-                        logger.debug("deleted old file: " + destFile.getAbsolutePath());
+                        System.out.print("This will delete file " + destFile.getAbsolutePath() + " are you sure y/N: ");
+                        try {
+                            if (Input.request().getAgreement()) {
+                                destFile.delete();
+                                logger.debug("deleted old file: " + destFile.getAbsolutePath());
+                            } else {
+                                continue;
+                            }
+                        } catch (InterruptedException e) {
+                            logger.debug("Interrupted while getting unarchive overwrite agreement");
+                            continue;
+                        }
                     }
                     file.renameTo(destFile);
                 }
